task_management/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dataclasses==0.6
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.7.0
kombu==5.4.2
prompt_toolkit==3.0.48
PyJWT==2.10.1
pymongo==4.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
